+++ cli/src/commands/fix.rs
@@ -27,7 +27,11 @@ use jj_lib::fileset::FilesetExpression;
 use jj_lib::fileset::FilesetParseContext;
 use jj_lib::fix::FileToFix;
 use jj_lib::fix::FixError;
+use jj_lib::fix::LineRange;
 use jj_lib::fix::ParallelFileFixer;
+use jj_lib::fix::RegionsToFormat;
+use jj_lib::fix::compute_changed_ranges;
+use jj_lib::fix::compute_file_line_count;
 use jj_lib::fix::fix_files;
 use jj_lib::matchers::Matcher;
 use jj_lib::repo::Repo as _;
@@ -170,6 +174,11 @@ pub(crate) struct FixArgs {
     /// specified, all files in the repo will be fixed.
     #[arg(long)]
     include_unchanged_files: bool,
+
+    /// Format all lines instead of only modified lines even if the formatter
+    /// supports formatting only modified lines.
+    #[arg(long, short)]
+    all_lines: bool,
 }
 
 #[instrument(skip_all)]
@@ -223,6 +232,7 @@ pub(crate) fn cmd_fix(
             &tools_config,
             store,
             file_to_fix,
+            args.all_lines,
         )
         .block_on()
     });
@@ -265,22 +275,70 @@ async fn fix_one_file(
     tools_config: &ToolsConfig,
     store: &Store,
     file_to_fix: &FileToFix,
+    all_lines_arg: bool,
 ) -> Result<Option<FileId>, FixError> {
     let mut matching_tools = tools_config
         .tools
         .iter()
         .filter(|tool_config| tool_config.matcher.matches(&file_to_fix.repo_path))
         .peekable();
+
     if matching_tools.peek().is_some() {
         // The first matching tool gets its input from the committed file, and any
         // subsequent matching tool gets its input from the previous matching tool's
         // output.
+
+        // TODO: Consider adding a check for some max file size config or some global
+        // limit for both `old_content` and `base_content`.
         let mut old_content = vec![];
         let mut read = store
             .read_file(&file_to_fix.repo_path, &file_to_fix.file_id)
             .await?;
         read.read_to_end(&mut old_content).await?;
+
+        // Do not run any tools on empty files.
+        if old_content.is_empty() {
+            return Ok(None);
+        }
+
+        // Load the base content from the file_to_fix (if exists) iff any tool needs it.
+        let base_content = match &file_to_fix.base_file_id {
+            Some(base_file_id) if matching_tools.clone().any(|t| t.line_range_arg.is_some()) => {
+                let mut content = vec![];
+                let mut read = store
+                    .read_file(&file_to_fix.repo_path, base_file_id)
+                    .await?;
+                read.read_to_end(&mut content).await?;
+                Some(content)
+            }
+            _ => None,
+        };
+
         let new_content = matching_tools.fold(old_content.clone(), |prev_content, tool_config| {
+            let mut extra_args = Vec::new();
+
+            if let Some(line_range_arg) = &tool_config.line_range_arg {
+                let ranges = match compute_regions_to_format(
+                    base_content.as_deref(),
+                    &prev_content,
+                    ComputeRegionsToFormatOptions {
+                        all_lines: all_lines_arg,
+                        skip_tool_if_zero_line_ranges: tool_config.skip_tool_if_zero_line_ranges,
+                    },
+                ) {
+                    RegionsToFormat::LineRanges(ranges) => ranges,
+                    RegionsToFormat::NoRegions => return prev_content,
+                };
+
+                for range in ranges {
+                    extra_args.push(
+                        line_range_arg
+                            .replace("$first", &range.first.to_string())
+                            .replace("$last", &range.last.to_string()),
+                    );
+                }
+            }
+
             match run_tool(
                 ui,
                 workspace_root,
@@ -288,6 +346,7 @@ async fn fix_one_file(
                 &tool_config.command,
                 file_to_fix,
                 &prev_content,
+                &extra_args,
             ) {
                 Ok(next_content) => next_content,
                 // TODO: Because the stderr is passed through, this isn't always failing
@@ -296,6 +355,7 @@ async fn fix_one_file(
                 Err(()) => prev_content,
             }
         });
+
         if new_content != old_content {
             // TODO: send futures back over channel
             let new_file_id = store
@@ -305,6 +365,53 @@ async fn fix_one_file(
         }
     }
     Ok(None)
+}
+
+/// Additional arguments for computing the modified line ranges between the base
+/// and current file.
+#[derive(Debug, Clone, Copy)]
+pub struct ComputeRegionsToFormatOptions {
+    /// Whether to compute the modified line ranges for all lines.
+    pub all_lines: bool,
+    /// Whether to skip the tool invocation if there are no line ranges to
+    /// format. For example, some tools want to sort imports even if there are
+    /// zero line ranges to format.
+    pub skip_tool_if_zero_line_ranges: bool,
+}
+
+/// Computes the modified line ranges between the base and current file.
+pub fn compute_regions_to_format(
+    base_content: Option<&[u8]>,
+    current_content: &[u8],
+    options: ComputeRegionsToFormatOptions,
+) -> RegionsToFormat {
+    if !options.all_lines
+        && let Some(base) = base_content
+    {
+        let changed_ranges = match compute_changed_ranges(base, current_content) {
+            RegionsToFormat::LineRanges(ranges) => ranges,
+            RegionsToFormat::NoRegions => return RegionsToFormat::NoRegions,
+        };
+
+        // If the tool is configured to skip the tool invocation if there are zero
+        // line ranges to format, we want to return NoRegions. Otherwise, we want to
+        // run the tool on the changed ranges.
+        if changed_ranges.is_empty() && options.skip_tool_if_zero_line_ranges {
+            return RegionsToFormat::NoRegions;
+        }
+        return RegionsToFormat::LineRanges(changed_ranges);
+    }
+
+    // Format the entire file.
+    let line_count = compute_file_line_count(current_content);
+    if line_count > 0 {
+        RegionsToFormat::LineRanges(vec![LineRange {
+            first: 1,
+            last: line_count,
+        }])
+    } else {
+        RegionsToFormat::NoRegions
+    }
 }
 
 /// Runs the `tool_command` to fix the given file content.
@@ -322,6 +429,7 @@ fn run_tool(
     tool_command: &CommandNameAndArgs,
     file_to_fix: &FileToFix,
     old_content: &[u8],
+    extra_args: &[String],
 ) -> Result<Vec<u8>, ()> {
     let mut vars: HashMap<&str, &str> = HashMap::new();
     vars.insert("path", file_to_fix.repo_path.as_internal_file_string());
@@ -333,6 +441,10 @@ fn run_tool(
         vars.insert("root", root);
     }
     let mut command = tool_command.to_command_with_variables(&vars);
+    if !extra_args.is_empty() {
+        command.args(extra_args);
+    }
+
     tracing::debug!(?command, ?file_to_fix.repo_path, "spawning fix tool");
     let Ok(mut child) = command
         .current_dir(workspace_root)
@@ -392,6 +504,20 @@ struct ToolConfig {
     matcher: Box<dyn Matcher>,
     /// Whether the tool is enabled
     enabled: bool,
+    /// Arguments to pass changed line ranges to the tool.
+    /// The string is a template where `$first` and `$last` are replaced by the
+    /// 1-based first and last inclusive line numbers of the changed range.
+    /// For example, `"--lines=$first-$last"`.
+    line_range_arg: Option<String>,
+    /// Whether to skip the tool invocation if there are zero line ranges to
+    /// format in the revision being fixed.
+    ///
+    /// This serves two main purposes:
+    /// - Avoiding unnecessary executions of tools. Set this to `false` if the
+    ///   tool should run regardless of diffs (e.g. to sort imports).
+    /// - Preventing tools configured with `line-range-arg` from making overly
+    ///   broad changes when zero line ranges are generated.
+    skip_tool_if_zero_line_ranges: bool,
     // TODO: Store the `name` field here and print it with the command's stderr, to clearly
     // associate any errors/warnings with the tool and its configuration entry.
 }
@@ -411,9 +537,17 @@ struct RawToolConfig {
     patterns: Vec<String>,
     #[serde(default = "default_tool_enabled")]
     enabled: bool,
+    #[serde(default)]
+    line_range_arg: Option<String>,
+    #[serde(default = "default_tool_skip_tool_if_zero_line_ranges")]
+    skip_tool_if_zero_line_ranges: bool,
 }
 
 fn default_tool_enabled() -> bool {
+    true
+}
+
+fn default_tool_skip_tool_if_zero_line_ranges() -> bool {
     true
 }
 
@@ -440,11 +574,21 @@ fn get_tools_config(
                     .map(|arg| fileset::parse(&mut diagnostics, arg, fileset_context))
                     .try_collect()?,
             );
+            if tool.line_range_arg.is_none()
+                && tool.skip_tool_if_zero_line_ranges
+                    != default_tool_skip_tool_if_zero_line_ranges()
+            {
+                return Err(config_error(
+                    "skip-tool-if-zero-line-ranges can only be set when line-range-arg is set",
+                ));
+            }
             print_parse_diagnostics(ui, &format!("In `fix.tools.{name}`"), &diagnostics)?;
             Ok(ToolConfig {
                 command: tool.command,
                 matcher: expression.to_matcher(),
                 enabled: tool.enabled,
+                line_range_arg: tool.line_range_arg,
+                skip_tool_if_zero_line_ranges: tool.skip_tool_if_zero_line_ranges,
             })
         })
         .try_collect()?;
@@ -458,5 +602,106 @@ fn get_tools_config(
         ))
     } else {
         Ok(ToolsConfig { tools })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_compute_regions_to_format_default() {
+        let options = ComputeRegionsToFormatOptions {
+            all_lines: false,
+            skip_tool_if_zero_line_ranges: true,
+        };
+
+        // Base content None.
+        assert_eq!(
+            compute_regions_to_format(None, b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..4).into()])
+        );
+
+        // Empty base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b""), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..4).into()])
+        );
+
+        // Modified base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nB\nc\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(2..3).into()])
+        );
+
+        // Deleted base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\nd\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::NoRegions
+        );
+
+        // Multiple line ranges.
+        assert_eq!(
+            compute_regions_to_format(Some(b"A\nb\nC\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..2).into(), (3..4).into()])
+        );
+
+        // Deleted current content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\n"), b"", options),
+            RegionsToFormat::NoRegions
+        );
+    }
+
+    #[test]
+    fn test_compute_regions_to_format_all_lines() {
+        let options = ComputeRegionsToFormatOptions {
+            all_lines: true,
+            skip_tool_if_zero_line_ranges: true,
+        };
+
+        // Empty base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b""), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..4).into()])
+        );
+
+        // Modified base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nB\nc\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..4).into()])
+        );
+
+        // Deleted base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\nd\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![(1..4).into()])
+        );
+
+        // Deleted current content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\n"), b"", options),
+            RegionsToFormat::NoRegions
+        );
+    }
+
+    #[test]
+    fn test_compute_regions_to_format_skip_tool_if_zero_line_ranges() {
+        let options = ComputeRegionsToFormatOptions {
+            all_lines: false,
+            skip_tool_if_zero_line_ranges: false,
+        };
+
+        // Deleted base content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\nd\n"), b"a\nb\nc\n", options),
+            RegionsToFormat::LineRanges(vec![])
+        );
+
+        // Deleted current content.
+        assert_eq!(
+            compute_regions_to_format(Some(b"a\nb\nc\n"), b"", options),
+            RegionsToFormat::NoRegions
+        );
     }
 }